+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/fragment/favorite/FavoriteViewModel.kt
@@ -1,0 +1,33 @@
+package com.sw.sw_api_kotlin_project.fragment.favorite
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.liveData
+import com.sw.sw_api_kotlin_project.base.BaseViewModel
+import com.sw.sw_api_kotlin_project.repository.FavoriteRepository
+import com.sw.sw_api_kotlin_project.utils.Resource
+import kotlinx.coroutines.Dispatchers
+
+class FavoriteViewModel(private val favoriteRepository: FavoriteRepository) : BaseViewModel() {
+
+    fun getFavoriteList() = liveData(Dispatchers.IO) {
+        try {
+            val response = favoriteRepository.getAll()
+            emit(Resource.success(response))
+        } catch (e: Exception) {
+            emit(Resource.error(data = null, message = e.message ?: "error"))
+        }
+    }
+
+}
+
+class FavoriteFactory(private val favoriteRepository: FavoriteRepository) :
+    ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(FavoriteViewModel::class.java)) {
+            @Suppress("UNCHECKED_CAST")
+            return FavoriteViewModel(favoriteRepository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
