--- SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/ui/home/HomeViewModel.kt
@@ -1,18 +1,0 @@
-package com.sw.sw_api_kotlin_project.ui.home
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.sw.sw_api_kotlin_project.base.BaseViewModel
-
-class HomeViewModel : BaseViewModel() {
-}
-
-class HomeFactory() : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(HomeViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return HomeViewModel() as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
