+++ SW_API_Kotlin_Project/app/src/main/res/layout/fragment_films_list.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="fragment_films"
-            type="com.sw.sw_api_kotlin_project.ui.films.FilmsListViewModel" />
+            type="com.sw.sw_api_kotlin_project.fragment.films.FilmsListViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
