+++ SW_API_Kotlin_Project/app/src/main/res/layout/fragment_others.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="fragment_others"
-            type="com.sw.sw_api_kotlin_project.ui.others.OthersViewModel" />
+            type="com.sw.sw_api_kotlin_project.fragment.others.OthersViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
